+++ include/Small/System.hpp
@@ -2,7 +2,7 @@
  * @file System.hpp
  * @author Alexandre Plateau (lexplt.dev@gmail.com)
  * @brief Include all the files from the System folder
- * @version 0.1
+ * @version 0.2
  * @date 2020-04-15
  * 
  * @copyright Copyright (c) 2020
@@ -12,7 +12,14 @@
 #ifndef sgl_small_system
 #define sgl_small_system
 
+/**
+ * @brief Functions and classes to related to system programming
+ * 
+ */
+namespace sgl::System {}
+
 #include <Small/System/Platform.hpp>
 #include <Small/System/SharedLibrary.hpp>
+#include <Small/System/FileSystem.hpp>
 
 #endif